rsoi-2022-lab1-ci-cd-malshevskikh-master/requirements.txt
Django==4.1.1
djangorestframework==3.14.0
psycopg2
psycopg2-binary==2.9.3